change hook useCharactersApi, simplify some functions, move all components from characters folder to components, replace colors with mui palette, work with constants
