+++ src/components/CustomTypography.jsx
@@ -1,0 +1,26 @@
+import React from 'react';
+import { Typography } from '@mui/material';
+
+const CustomTypography = (props) => {
+  return (
+    <Typography
+      variant='h1'
+      sx={{
+        fontSize: {
+          xs: '2rem', // small screens
+          sm: '2.5rem', // medium screens
+          md: '3rem', // large screens
+          lg: '3.5rem', // extra large screens
+          xl: '4rem', // double extra large screens
+        },
+        marginBottom: {
+          xs: '20px',
+          md: '30px',
+        },
+      }}
+      {...props}
+    />
+  );
+};
+
+export default CustomTypography;
