+++ src/components/GenderRadioButtons.jsx
@@ -6,24 +6,28 @@ import {
   RadioGroup,
   Radio,
 } from '@mui/material';
+import { useTheme } from '@mui/material/styles';
 
-import { useCharactersContext } from '../../context/CharactersContext';
-import { commonStyles, GENDER_OPTIONS } from '../../constants';
+import { useCharactersContext } from '../context/CharactersContext';
+import { FILTER_NAMES, GENDER_OPTIONS } from '../constants';
 
 const GenderRadioButtons = () => {
   const { filters, handleFilterChange } = useCharactersContext();
 
+  const { palette } = useTheme();
+
   const handleGenderChange = (e) =>
-    handleFilterChange('gender', e.target.value);
+    handleFilterChange(FILTER_NAMES.gender, e.target.value);
 
   return (
     <FormControl>
-      <FormLabel sx={{ color: commonStyles.secondaryTextColor }}>
-        Gender
+      <FormLabel sx={{ color: palette.grey[400] }}>
+        {FILTER_NAMES.gender.charAt(0).toUpperCase() +
+          FILTER_NAMES.gender.slice(1)}
       </FormLabel>
       <RadioGroup
         row
-        name='gender'
+        name={FILTER_NAMES.gender}
         value={filters.gender}
         onChange={handleGenderChange}
       >
@@ -31,7 +35,7 @@ const GenderRadioButtons = () => {
           <FormControlLabel
             key={option}
             value={option}
-            control={<Radio sx={{ color: commonStyles.secondaryTextColor }} />}
+            control={<Radio sx={{ color: palette.grey[400] }} />}
             label={option.charAt(0).toUpperCase() + option.slice(1)}
           />
         ))}
