+++ src/components/StatusRadioButtons.jsx
@@ -6,24 +6,28 @@ import {
   RadioGroup,
   Radio,
 } from '@mui/material';
+import { useTheme } from '@mui/material/styles';
 
-import { useCharactersContext } from '../../context/CharactersContext';
-import { commonStyles, STATUS_OPTIONS } from '../../constants';
+import { useCharactersContext } from '../context/CharactersContext';
+import { FILTER_NAMES, STATUS_OPTIONS } from '../constants';
 
 const StatusRadioButtons = () => {
   const { filters, handleFilterChange } = useCharactersContext();
 
+  const { palette } = useTheme();
+
   const handleStatusChange = (e) =>
-    handleFilterChange('status', e.target.value);
+    handleFilterChange(FILTER_NAMES.status, e.target.value);
 
   return (
     <FormControl>
-      <FormLabel sx={{ color: commonStyles.secondaryTextColor }}>
-        Status
+      <FormLabel sx={{ color: palette.grey[400] }}>
+        {FILTER_NAMES.status.charAt(0).toUpperCase() +
+          FILTER_NAMES.status.slice(1)}
       </FormLabel>
       <RadioGroup
         row
-        name='status'
+        name={FILTER_NAMES.status}
         value={filters.status}
         onChange={handleStatusChange}
       >
@@ -31,7 +35,7 @@ const StatusRadioButtons = () => {
           <FormControlLabel
             key={option}
             value={option}
-            control={<Radio sx={{ color: commonStyles.secondaryTextColor }} />}
+            control={<Radio sx={{ color: palette.grey[400] }} />}
             label={option.charAt(0).toUpperCase() + option.slice(1)}
           />
         ))}
