--- src/components/characters/typographyStyles.js
@@ -1,33 +1,0 @@
-export const h1Styled = {
-  fontSize: {
-    xs: '2rem', // small screens
-    sm: '2.5rem', // medium screens
-    md: '3rem', // large screens
-    lg: '3.5rem', // extra large screens
-    xl: '4rem', // double extra large screens
-  },
-  margin: {
-    xs: '160px 0 5px 0',
-    sm: '100px 0 20px 0',
-    md: '100px 0 20px 0',
-    lg: '110px 0 20px 0',
-    xl: '110px 0 20px 0',
-  },
-};
-
-export const h2Styled = {
-  fontSize: {
-    xs: '2rem', // small screens
-    sm: '2.4rem', // medium screens
-    md: '2.8rem', // large screens
-    lg: '3.2rem', // extra large screens
-    xl: '3.6rem', // double extra large screens
-  },
-  margin: {
-    xs: '30px 0 20px 0',
-    sm: '160px 0 20px 0',
-    md: '130px 0 30px 0',
-    lg: '90px 0 30px 0',
-    xl: '80px 0 30px 0',
-  },
-};
