+++ src/components/index.js
@@ -9,8 +9,6 @@ import GenderRadioButtons from './GenderRadioButtons';
 import ErrorDisplay from './ErrorDisplay';
 import LoadingDisplay from './LoadingDisplay';
 
-import { h1Styled, h2Styled } from './typographyStyles';
-
 export {
   Card,
   Cards,
@@ -22,6 +20,4 @@ export {
   GenderRadioButtons,
   ErrorDisplay,
   LoadingDisplay,
-  h1Styled,
-  h2Styled,
 };
