+++ src/constants.js
@@ -1,12 +1,9 @@
-export const commonStyles = {
-  backgroundColor: '#282c34',
-  primaryTextColor: 'white',
-  secondaryTextColor: '#ccc',
-  borderColor: 'white',
-  linkColor: '#1976d2',
-};
-
 export const CHARACTER_API_URL = 'https://rickandmortyapi.com/api/character/';
 export const ALL_SPECIES_NAME = 'All Species';
+export const FILTER_NAMES = {
+  name: 'name',
+  status: 'status',
+  gender: 'gender',
+};
 export const GENDER_OPTIONS = ['male', 'female', 'genderless', 'unknown'];
 export const STATUS_OPTIONS = ['alive', 'dead', 'unknown'];
